refactor `InputHandler`
