+++ source/input/InputService.ts
@@ -1,9 +1,10 @@
 import process from "node:process";
-import { EventEmitter } from "#events";
+
+export type InputHandler = (data: Buffer) => void;
 
 export interface ReadStream {
-  addListener: (event: "data", listener: (data: Buffer) => void) => this;
-  removeListener: (event: "data", listener: (data: Buffer) => void) => this;
+  addListener: (event: "data", handler: InputHandler) => this;
+  removeListener: (event: "data", handler: InputHandler) => this;
   setRawMode?: (mode: boolean) => this;
   unref: () => this;
 }
@@ -13,9 +14,11 @@ export interface InputServiceOptions {
 }
 
 export class InputService {
+  #onKeyPressed: InputHandler;
   #stdin: ReadStream;
 
-  constructor(options?: InputServiceOptions) {
+  constructor(onKeyPressed: InputHandler, options?: InputServiceOptions) {
+    this.#onKeyPressed = onKeyPressed;
     this.#stdin = options?.stdin ?? process.stdin;
 
     this.#stdin.setRawMode?.(true);
@@ -26,9 +29,5 @@ export class InputService {
 
   close(): void {
     this.#stdin.removeListener("data", this.#onKeyPressed);
-  }
-
-  #onKeyPressed(this: void, data: Buffer): void {
-    EventEmitter.dispatch(["input:info", { key: data.toString() }]);
   }
 }
