+++ source/input/index.ts
@@ -1,1 +1,1 @@
-export { InputService, type InputServiceOptions, type ReadStream } from "./InputService.js";
+export { InputService, type InputHandler, type InputServiceOptions, type ReadStream } from "./InputService.js";
